change log funcs files
